testing framework: move local implementation into local directory (#33734)
