+++ internal/backend/local/backend_local.go
@@ -10,6 +10,9 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/hashicorp/hcl/v2"
+	"github.com/zclconf/go-cty/cty"
+
 	"github.com/hashicorp/terraform/internal/backend"
 	"github.com/hashicorp/terraform/internal/configs"
 	"github.com/hashicorp/terraform/internal/configs/configload"
@@ -17,7 +20,6 @@ import (
 	"github.com/hashicorp/terraform/internal/states/statemgr"
 	"github.com/hashicorp/terraform/internal/terraform"
 	"github.com/hashicorp/terraform/internal/tfdiags"
-	"github.com/zclconf/go-cty/cty"
 )
 
 // backend.Local implementation.
@@ -503,4 +505,23 @@ func (v unparsedUnknownVariableValue) ParseVariableValue(mode configs.VariablePa
 		Value:      cty.UnknownVal(v.WantType),
 		SourceType: terraform.ValueFromInput,
 	}, nil
+}
+
+type unparsedTestVariableValue struct {
+	expr hcl.Expression
+	ctx  *hcl.EvalContext
+}
+
+func (v unparsedTestVariableValue) ParseVariableValue(mode configs.VariableParsingMode) (*terraform.InputValue, tfdiags.Diagnostics) {
+	var diags tfdiags.Diagnostics
+	val, hclDiags := v.expr.Value(v.ctx) // nil because no function calls or variable references are allowed here
+	diags = diags.Append(hclDiags)
+
+	rng := tfdiags.SourceRangeFromHCL(v.expr.Range())
+
+	return &terraform.InputValue{
+		Value:       val,
+		SourceType:  terraform.ValueFromConfig, // Test variables always come from config.
+		SourceRange: rng,
+	}, diags
 }
