+++ internal/moduletest/suite.go
@@ -3,8 +3,16 @@
 
 package moduletest
 
+import "github.com/hashicorp/terraform/internal/tfdiags"
+
 type Suite struct {
 	Status Status
 
 	Files map[string]*File
+}
+
+type TestSuiteRunner interface {
+	Test() (Status, tfdiags.Diagnostics)
+	Stop()
+	Cancel()
 }
